Restructure to put logic in lib
